+++ tasks/taurus/src/main/filtered-resources/com/walmartlabs/concord/plugins/taurus/DummyPluginsManagerCMD.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+# a fake PluginsManagerCMD to prevent the task from accessing the JMeter plugin registry
+exit 0